+++ lib/posix-tty/pseudo.c
@@ -23,7 +23,7 @@ static const char ZERO_VOLID[] = "zero_vol";
 
 static ssize_t null_read(const struct uk_file *f __maybe_unused,
 			 const struct iovec *iov __unused, int iovcnt __unused,
-			 off_t off __unused, long flags __unused)
+			 size_t off __unused, long flags __unused)
 {
 	UK_ASSERT(f->vol == NULL_VOLID);
 	return 0;
@@ -31,7 +31,7 @@ static ssize_t null_read(const struct uk_file *f __maybe_unused,
 
 static ssize_t void_read(const struct uk_file *f __maybe_unused,
 			 const struct iovec *iov __unused, int iovcnt __unused,
-			 off_t off __unused, long flags __unused)
+			 size_t off __unused, long flags __unused)
 {
 	UK_ASSERT(f->vol == VOID_VOLID);
 	return -EAGAIN;
@@ -39,7 +39,7 @@ static ssize_t void_read(const struct uk_file *f __maybe_unused,
 
 static ssize_t zero_read(const struct uk_file *f __maybe_unused,
 			 const struct iovec *iov, int iovcnt,
-			 off_t off __unused, long flags __unused)
+			 size_t off __unused, long flags __unused)
 {
 	ssize_t total = 0;
 
@@ -56,7 +56,7 @@ static ssize_t zero_read(const struct uk_file *f __maybe_unused,
 
 static ssize_t null_write(const struct uk_file *f __maybe_unused,
 			  const struct iovec *iov, int iovcnt,
-			  off_t off __unused, long flags __unused)
+			  size_t off __unused, long flags __unused)
 {
 	ssize_t total = 0;
