+++ lib/ukfile/file-nops.c
@@ -11,14 +11,14 @@
 
 ssize_t uk_file_nop_read(const struct uk_file *f __unused,
 			 const struct iovec *iov __unused, int iovcnt __unused,
-			 off_t off __unused, long flags __unused)
+			 size_t off __unused, long flags __unused)
 {
 	return -ENOSYS;
 }
 
 ssize_t uk_file_nop_write(const struct uk_file *f __unused,
 			  const struct iovec *iov __unused, int iovcnt __unused,
-			  off_t off __unused, long flags __unused)
+			  size_t off __unused, long flags __unused)
 {
 	return -ENOSYS;
 }
